Make timestamp logging more portable

+++ src/main.cpp
@@ -121,9 +121,9 @@ void thd_logger(const gchar *log_domain, GLogLevelFlags log_level,
 	seconds = time(NULL);
 
 	if (thd_daemonize)
-		syslog(syslog_priority, "[%ld]%s%s", seconds, prefix, message);
+		syslog(syslog_priority, "[%lld]%s%s", (long long) seconds, prefix, message);
 	else
-		g_print("[%ld]%s%s", seconds, prefix, message);
+		g_print("[%lld]%s%s", (long long) seconds, prefix, message);
 
 }
